Refactor API models to separate files
Moved API model data classes from ApiModels.kt
to individual files under the `model.api` package.
Updated all usages to reflect the new package structure.
