+++ app/src/main/java/io/github/deprec8/enigmadroid/data/ApiRepository.kt
@@ -31,18 +31,18 @@ import io.github.deprec8.enigmadroid.data.objects.PreferencesKeys
 import io.github.deprec8.enigmadroid.data.source.local.devices.Device
 import io.github.deprec8.enigmadroid.data.source.local.devices.DevicesDatabase
 import io.github.deprec8.enigmadroid.data.source.network.NetworkDataSource
-import io.github.deprec8.enigmadroid.model.Bookmark
-import io.github.deprec8.enigmadroid.model.BookmarkList
-import io.github.deprec8.enigmadroid.model.BouquetList
-import io.github.deprec8.enigmadroid.model.CurrentInfo
-import io.github.deprec8.enigmadroid.model.DeviceInfo
-import io.github.deprec8.enigmadroid.model.EPGEventList
-import io.github.deprec8.enigmadroid.model.EventList
-import io.github.deprec8.enigmadroid.model.MovieList
-import io.github.deprec8.enigmadroid.model.ServiceList
-import io.github.deprec8.enigmadroid.model.SignalInfo
-import io.github.deprec8.enigmadroid.model.Timer
-import io.github.deprec8.enigmadroid.model.TimerList
+import io.github.deprec8.enigmadroid.model.api.Bookmark
+import io.github.deprec8.enigmadroid.model.api.BookmarkList
+import io.github.deprec8.enigmadroid.model.api.BouquetList
+import io.github.deprec8.enigmadroid.model.api.CurrentInfo
+import io.github.deprec8.enigmadroid.model.api.DeviceInfo
+import io.github.deprec8.enigmadroid.model.api.EPGEventList
+import io.github.deprec8.enigmadroid.model.api.EventList
+import io.github.deprec8.enigmadroid.model.api.MovieList
+import io.github.deprec8.enigmadroid.model.api.ServiceList
+import io.github.deprec8.enigmadroid.model.api.SignalInfo
+import io.github.deprec8.enigmadroid.model.api.Timer
+import io.github.deprec8.enigmadroid.model.api.TimerList
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.asFlow
