+++ app/src/main/java/io/github/deprec8/enigmadroid/data/DownloadRepository.kt
@@ -30,7 +30,7 @@ import io.github.deprec8.enigmadroid.R
 import io.github.deprec8.enigmadroid.data.objects.PreferencesKeys
 import io.github.deprec8.enigmadroid.data.source.local.devices.Device
 import io.github.deprec8.enigmadroid.data.source.local.devices.DevicesDatabase
-import io.github.deprec8.enigmadroid.model.Movie
+import io.github.deprec8.enigmadroid.model.api.Movie
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.flow.map
